Consolidated Win32 changes to compatibility header file, and cleaned up several #ifdefs
