+++ compat/win32/win32_clock_gettime.c
