+++ dump1090.h
@@ -71,76 +71,17 @@
 #include <signal.h>
 #include <fcntl.h>
 #include <ctype.h>
-#include <sys/stat.h>
 #include <time.h>
 #include <limits.h>
-
-#define CONCAT(A, B) A ## B
 
 #ifndef _WIN32
 
 #include <unistd.h>
-#include <sys/time.h>
 #include <sys/ioctl.h>
+#include <sys/stat.h>
+#include <sys/time.h>
 
-#define OPEN_FLAGS O_RDONLY
-
-#define sockErrorIs(A) (errno == CONCAT(E, A))
-
-#else
-
-#define __attribute__(A)
-
-#include <io.h>
-#include <winsock2.h>
-#include <windows.h>
-
-#include "compat/win32_clock_nanosleep.h"
-
-typedef intptr_t ssize_t;
-
-#define M_PI 3.14159265358979323846264338327950288
-#define STDIN_FILENO _fileno(stdin)
-#define STDOUT_FILENO _fileno(stdout)
-
-#ifndef PATH_MAX
-#define PATH_MAX 260
-#endif
-
-#define OPEN_FLAGS (O_RDONLY | O_BINARY)
-
-#define sockErrorIs(A) (WSAGetLastError() == CONCAT(WSAE, A))
-
-#define realpath(P, R) strcpy(R, P)
-
-static inline struct tm *localtime_r(const time_t *timep, struct tm *result)
-{
-    localtime_s(result, timep);
-    return result;
-}
-
-static inline int usleep(__int64 usec)
-{
-    struct timespec ts;
-
-    ts.tv_sec = (time_t)usec / 1000000;
-    ts.tv_nsec = (long)(usec % 1000000) * 1000;
-
-    int err = clock_nanosleep(CLOCK_MONOTONIC, 0, &ts, NULL);
-
-    if (err != 0)
-    {
-        errno = err;
-        return -1;
-    }
-
-    return 0;
-}
-
-#define sleep(sec) usleep((sec) * 1000000ULL)
-#define strcasecmp stricmp
-
-#endif
+#endif // _WIN32
 
 #include "compat/compat.h"
