+++ public_html/registrations.js
